Do not even add the testing executables when the tests are disabled.
I think this should be a more complex option to allow every possibility which was available previously, but for now, it'll do it for me:

* If I want to build the test suite, it'll build it
* If I don't need it (most of the time), it'll completely ignore test targets

Focus is on the second point, as most of the CMake configuration time was spent on the test targets. With this change, build file generation only takes a few seconds.